+++ Giraf.IntegrationTests/Utils/DbSeeders/EmptyDb.cs
@@ -6,6 +6,6 @@ public class EmptyDb : DbSeeder
 {
     public override void SeedData(DbContext context)
     {
-        context.Database.EnsureDeleted();
+
     }
 }
